Fix resetting of clock count
Instead of resetting the clock count to 0, subtract by 1024 (which is
the highest needed clock count for timers).
This is change is needed because an instruction can take longer than 4
opcodes and therefore the timers can become inaccurate after a reset.

Example: counter is at 1020 and an opcode gets executed which takes 8
opcodes. If we reset to 0 we won't be taking the 4 extra cycles into
account.